HomeFragment data and domain module added

+++ shared/accounts/data/src/main/java/com/chelz/shared/accounts/domain/dto/CategoryDto.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "Categories")
 data class CategoryDto(
-	@PrimaryKey(autoGenerate = true) val categoryId: Int,
+	@PrimaryKey(autoGenerate = true) val categoryId: Long,
 	@ColumnInfo(name = "name") var name: String,
 	@ColumnInfo(name = "isEarning") var isEarning: Boolean = false,
 	@ColumnInfo(name = "color") var color: String = "#FF0000",
